feat: add form progress bar
