+++ src/app/admin/dispatch/page.tsx
@@ -5,6 +5,7 @@ import { useRouter } from 'next/navigation';
 import AdminAuthGuard from '@/components/AdminAuthGuard';
 import TruckRegistration from '@/components/TruckRegistration';
 import DispatchCalendar from '@/components/DispatchCalendar';
+import TruckAssignmentModal from './components/TruckAssignmentModal';
 
 interface Truck {
   id: string;
@@ -575,220 +576,6 @@ export default function DispatchManagement() {
   };
 
   // トラック割り当てモーダル
-  const TruckAssignmentModal = () => {
-    const [modalSelectedTruck, setModalSelectedTruck] = useState<any>(null);
-    const [formData, setFormData] = useState({
-      capacity: '',
-      startTime: '09:00',
-      endTime: '17:00',
-      workType: 'loading' as 'loading' | 'moving' | 'unloading',
-    });
-
-    useEffect(() => {
-      if (selectedSubmission) {
-        setFormData({
-          capacity: '',
-          startTime: '09:00',
-          endTime: '17:00',
-          workType: 'loading',
-        });
-        setModalSelectedTruck(null);
-      }
-    }, [selectedSubmission]);
-
-    // 推奨トラックを計算
-    const recommendedTrucks = selectedSubmission ? 
-      calculateRecommendedTrucks(selectedSubmission.totalPoints) : [];
-
-    // 見積もり価格を計算
-    const estimatedPrice = selectedSubmission ? 
-      calculateEstimatedPrice(selectedSubmission.totalPoints, selectedSubmission.distance || 0) : 0;
-
-    const handleSubmit = (e: React.FormEvent) => {
-      e.preventDefault();
-      
-      if (!modalSelectedTruck || !selectedSubmission) return;
-
-      const truckAssignment: TruckAssignment = {
-        truckId: modalSelectedTruck.id,
-        truckName: modalSelectedTruck.name,
-        capacity: parseInt(formData.capacity),
-        startTime: formData.startTime,
-        endTime: formData.endTime,
-        workType: formData.workType,
-      };
-
-      assignTruckToSubmission(selectedSubmission.id, truckAssignment);
-      setShowTruckModal(false);
-      setModalSelectedTruck(null);
-    };
-
-    // 料金設定のトラックから利用可能なトラックをフィルタリング
-    const availablePricingTrucks = pricingTrucks.filter(truck => {
-      // 指定日時にスケジュールが重複していないかチェック（配車管理のトラックスケジュールと照合）
-      const hasConflict = trucks.some(dispatchTruck => 
-        dispatchTruck.schedules.some(schedule => 
-          schedule.date === selectedSubmission?.moveDate &&
-          schedule.status === 'booked' &&
-          ((schedule.startTime <= formData.startTime && schedule.endTime > formData.startTime) ||
-           (schedule.startTime < formData.endTime && schedule.endTime >= formData.endTime) ||
-           (schedule.startTime >= formData.startTime && schedule.endTime <= formData.endTime))
-        )
-      );
-      
-      return truck.status === 'available' && !hasConflict;
-    });
-
-    return (
-      <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-        <div className="bg-white rounded-lg p-6 w-full max-w-2xl max-h-[90vh] overflow-y-auto">
-          <h3 className="text-lg font-semibold mb-4">トラック割り当て</h3>
-          
-          {selectedSubmission && (
-            <div className="mb-4 p-4 bg-blue-50 rounded">
-              <h4 className="font-medium text-blue-900">{selectedSubmission.customerName}</h4>
-              <p className="text-sm text-blue-700">{selectedSubmission.customerEmail}</p>
-              <p className="text-sm text-blue-600">
-                {formatDate(selectedSubmission.moveDate)} {formatTime(formData.startTime)}-{formatTime(formData.endTime)}
-              </p>
-              <div className="grid grid-cols-2 gap-4 mt-2">
-                <div>
-                  <p className="text-sm text-blue-600">
-                    総容量: {selectedSubmission.totalCapacity.toLocaleString()}kg
-                  </p>
-                  <p className="text-sm text-blue-600">
-                    総ポイント: {selectedSubmission.totalPoints}pt
-                  </p>
-                </div>
-                <div>
-                  <p className="text-sm text-blue-600">
-                    見積もり価格: ¥{estimatedPrice.toLocaleString()}
-                  </p>
-                  {selectedSubmission.distance && (
-                    <p className="text-sm text-blue-600">
-                      距離: {selectedSubmission.distance}km
-                    </p>
-                  )}
-                </div>
-              </div>
-              
-              {/* 推奨トラック */}
-              {recommendedTrucks.length > 0 && (
-                <div className="mt-3">
-                  <p className="text-sm font-medium text-blue-900 mb-2">推奨トラック:</p>
-                  <div className="flex flex-wrap gap-2">
-                    {recommendedTrucks.map(truck => (
-                      <span
-                        key={truck.id}
-                        className={`px-2 py-1 text-xs rounded cursor-pointer ${
-                          modalSelectedTruck?.id === truck.id
-                            ? 'bg-blue-600 text-white'
-                            : 'bg-blue-100 text-blue-800 hover:bg-blue-200'
-                        }`}
-                        onClick={() => setModalSelectedTruck(truck)}
-                      >
-                        {truck.name} ({truck.truckType})
-                      </span>
-                    ))}
-                  </div>
-                </div>
-              )}
-            </div>
-          )}
-          
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div>
-              <label className="block text-sm font-medium mb-1">トラック選択</label>
-              <select
-                value={modalSelectedTruck?.id || ''}
-                onChange={e => {
-                  const truck = availablePricingTrucks.find(t => t.id === e.target.value);
-                  setModalSelectedTruck(truck || null);
-                }}
-                className="w-full px-3 py-2 border rounded"
-                required
-              >
-                <option value="">トラックを選択</option>
-                {availablePricingTrucks.map(truck => (
-                  <option key={truck.id} value={truck.id}>
-                    {truck.name} ({truck.plateNumber}) - {truck.truckType} - {truck.capacityKg}kg - ¥{truck.basePrice.toLocaleString()}
-                  </option>
-                ))}
-              </select>
-            </div>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div>
-                <label className="block text-sm font-medium mb-1">割り当て容量（kg）</label>
-                <input
-                  type="number"
-                  value={formData.capacity}
-                  onChange={e => setFormData({ ...formData, capacity: e.target.value })}
-                  className="w-full px-3 py-2 border rounded"
-                  required
-                  min="0"
-                  max={modalSelectedTruck?.capacityKg}
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">作業区分</label>
-                <select
-                  value={formData.workType}
-                  onChange={e => setFormData({ ...formData, workType: e.target.value as any })}
-                  className="w-full px-3 py-2 border rounded"
-                >
-                  <option value="loading">積込</option>
-                  <option value="moving">移動</option>
-                  <option value="unloading">積卸</option>
-                </select>
-              </div>
-            </div>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div>
-                <label className="block text-sm font-medium mb-1">開始時間</label>
-                <input
-                  type="time"
-                  value={formData.startTime}
-                  onChange={e => setFormData({ ...formData, startTime: e.target.value })}
-                  className="w-full px-3 py-2 border rounded"
-                  required
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">終了時間</label>
-                <input
-                  type="time"
-                  value={formData.endTime}
-                  onChange={e => setFormData({ ...formData, endTime: e.target.value })}
-                  className="w-full px-3 py-2 border rounded"
-                  required
-                />
-              </div>
-            </div>
-            
-            <div className="flex gap-2 pt-4">
-              <button
-                type="submit"
-                className="flex-1 bg-blue-600 text-white py-2 rounded hover:bg-blue-700"
-              >
-                割り当て
-              </button>
-              <button
-                type="button"
-                onClick={() => {
-                  setShowTruckModal(false);
-                  setModalSelectedTruck(null);
-                }}
-                className="px-4 py-2 bg-gray-600 text-white rounded hover:bg-gray-700"
-              >
-                キャンセル
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-    );
   };
 
   const handleLogout = () => {
@@ -1062,7 +849,17 @@ export default function DispatchManagement() {
         </main>
 
         {/* モーダル */}
-        {showTruckModal && <TruckAssignmentModal />}
+        {showTruckModal && (
+          <TruckAssignmentModal
+            selectedSubmission={selectedSubmission}
+            trucks={trucks}
+            pricingTrucks={pricingTrucks}
+            setShowTruckModal={setShowTruckModal}
+            assignTruckToSubmission={assignTruckToSubmission}
+            calculateRecommendedTrucks={calculateRecommendedTrucks}
+            calculateEstimatedPrice={calculateEstimatedPrice}
+          />
+        )}
       </div>
     </AdminAuthGuard>
   );
