+++ src/app/form/step2/page.tsx
@@ -6,6 +6,8 @@
 import { useForm } from 'react-hook-form';
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
+import ProgressBar from '@/components/ProgressBar';
+import { ITEM_CATEGORY_NAMES } from '@/constants/items';
 
 // 定数定義
 const AUTOSAVE_INTERVAL = 5000;
@@ -23,141 +25,6 @@ const styles = {
   }
 };
 
-// 荷物データ定義
-const itemCategories = [
-  {
-    category: "ベッド",
-    data: [
-      "🛏️ シングルベッド",
-      "🛏️ セミダブルベッド",
-      "🛏️ ダブルベッド",
-      "🛏️ クイーンベッド",
-      "🛏️ キングベッド",
-      "🛏️‍🛏️ 2段ベッド",
-      "👶 ベビーベッド"
-    ]
-  },
-  {
-    category: "マットレス・布団",
-    data: [
-      "🛏️ マットレス（シングル）",
-      "🛏️ マットレス（セミダブル）",
-      "🛏️ マットレス（ダブル）",
-      "🛏️ マットレス（クイーン）",
-      "🛏️ マットレス（キング）",
-      "🛌 布団類（羽毛、毛布など）"
-    ]
-  },
-  {
-    category: "ソファ",
-    data: [
-      "🛋️ ソファ（1人掛け）",
-      "🛋️ ソファ（2人掛け）",
-      "🛋️ ソファ（3人掛け）",
-      "🛋️ カウチソファ",
-      "🛋️ ソファベッド",
-      "🛋️ オットマン"
-    ]
-  },
-  {
-    category: "衣類収納",
-    data: [
-      "👕 ハンガーラック",
-      "🎽 衣装ケース",
-      "🚪 洋服タンス"
-    ]
-  },
-  {
-    category: "棚・収納",
-    data: [
-      "🗄 タンス（小型）",
-      "🗄 タンス（大型）",
-      "📚 本棚",
-      "🍽 食器棚",
-      "🖥 テレビ台",
-      "🧸 おもちゃ棚・キッズ収納",
-      "📦 カラーボックス"
-    ]
-  },
-  {
-    category: "テーブル・机類",
-    data: [
-      "🍴 ダイニングテーブル（2人用）",
-      "🍴 ダイニングテーブル（4人用）",
-      "🍴 ダイニングテーブル（6人以上）",
-      "🛋 リビングテーブル",
-      "🛋 ローテーブル",
-      "🪵 こたつ",
-      "💄 ドレッサー",
-      "💻 パソコンデスク",
-      "🪑 チェア・イス"
-    ]
-  },
-  {
-    category: "家電",
-    data: [
-      "🧺 洗濯機（縦型）",
-      "🧺 洗濯機（ドラム式）",
-      "🧊 冷蔵庫（小型）",
-      "🧊 冷蔵庫（大型）",
-      "📺 テレビ（40インチ未満）",
-      "📺 テレビ（40〜60インチ）",
-      "📺 テレビ（60インチ以上）",
-      "💻 パソコン（ノート）",
-      "💻 パソコン（デスクトップ）",
-      "🍱 電子レンジ",
-      "🍞 オーブントースター",
-      "🍚 炊飯器",
-      "🔥 ストーブ・ヒーター",
-      "❄️ エアコン（本体＋室外機）",
-      "📡 掃除機",
-      "🧼 加湿器 / 空気清浄機",
-      "🤖 ロボット掃除機"
-    ]
-  },
-  {
-    category: "特殊・大型アイテム",
-    data: [
-      "🚲 自転車",
-      "🏍 バイク",
-      "🎹 ピアノ（アップライト）",
-      "🎹 ピアノ（グランド）",
-      "🎹 電子ピアノ",
-      "🔐 金庫（小型）",
-      "🔐 金庫（大型）",
-      "🐠 水槽（30cm以下）",
-      "🐠 水槽（30cm以上）"
-    ]
-  },
-  {
-    category: "生活雑貨",
-    data: [
-      "🪴 観葉植物（小型）",
-      "🪴 観葉植物（大型）",
-      "🧳 キャリーケース",
-      "🪞 姿見・鏡",
-      "🗑 ゴミ箱（大型）",
-      "🪣 バケツ・掃除道具セット",
-      "🧺 ランドリーバスケット",
-      "🏠 物干し竿"
-    ]
-  },
-  {
-    category: "その他家電・日用品",
-    data: [
-      "🖨 プリンター",
-      "🔊 ホームシアター（本体のみ）",
-      "🎥 スクリーン",
-      "🔈 スピーカー",
-      "📻 アンプ",
-      "🎮 ゲーム機",
-      "🍳 ホットプレート",
-      "💡 シーリングライト",
-      "🪞 スタンドライト",
-      "🪟 カーテン（左右セット）"
-    ]
-  }
-];
 
 const boxSizeOptions = [
   "🏠 10箱未満（荷物が少ない）",
@@ -167,13 +34,18 @@ const boxSizeOptions = [
   "🏠 51箱以上（3LDK / 4K以上の目安）"
 ];
 
+interface Step2FormData {
+  items: Record<string, any>;
+  itemsRemarks?: string;
+}
+
 export default function Step2FormPage() {
-  const { register, handleSubmit, setValue, watch } = useForm();
+  const { register, handleSubmit, setValue, watch } = useForm<Step2FormData>();
   const router = useRouter();
   const danball = watch('items.danball');
 
   // フォームデータの保存
-  const saveFormData = (data: any) => {
+  const saveFormData = (data: Step2FormData) => {
     try {
       localStorage.setItem(STORAGE_KEY, JSON.stringify(data));
     } catch (e) {
@@ -182,7 +54,7 @@ export default function Step2FormPage() {
   };
 
   // フォーム送信処理
-  const onSubmit = (data: any) => {
+  const onSubmit = (data: Step2FormData) => {
     try {
       saveFormData(data);
       router.push('/form/step3');
@@ -221,13 +93,14 @@ export default function Step2FormPage() {
     <main className="bg-gray-50 min-h-screen py-10 px-4">
       <form onSubmit={handleSubmit(onSubmit)} className="max-w-4xl mx-auto space-y-10 text-gray-800">
         <h1 className="text-3xl font-bold text-center text-blue-800">📦 荷物の数量を入力</h1>
+        <ProgressBar current={2} total={3} />
         <div>
           <p className="text-center text-sm text-gray-600">必要なものをすべて入力してください（0でもOK）</p>
           <p className="text-center text-sm text-gray-500">※ 段ボールに梱包できるものは入力不要です</p>
         </div>
 
         {/* 家具・家電の数量入力 */}
-        {itemCategories.map(({ category, data }) => (
+        {ITEM_CATEGORY_NAMES.map(({ category, data }) => (
           <section key={category} className={styles.section}>
             <h2 className="text-lg font-semibold mb-4">🗂 {category}</h2>
             <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
